app/twitter_scraper.py
# app/twitter_scraper.py

import requests
from bs4 import BeautifulSoup

def scrape_twitter(term: str):
    # Mocked URL and response for illustration; replace with actual Twitter scraping logic
    url = f"https://twitter.com/search?q={term}"
    response = requests.get(url)
    soup = BeautifulSoup(response.text, 'html.parser')
    tweets = [tweet.text for tweet in soup.find_all('div', class_='tweet')]
    return tweets
