requirements.txt
fastapi
sqlalchemy
psycopg2-binary
openai
requests
beautifulsoup4
uvicorn
python-dotenv
